requirements.txt
httpx~=0.28.1
aiohttp~=3.11.11
python-dotenv~=1.0.1
TonTools~=2.1.2